Changing the signature of \cyclone\db\prepared\Executor: adding the original query to all exec_* methods as parameter
A third parameter is added to the following methods of \cyclone\db\prepared\Executor:
* exec_insert()
* exec_update()
* exec_delete()
The parameter is needed by the executing methods to retrieve the the columns listed in the RETURNING clause
of the original SQL query. In the case of MySQL these are needed for the SELECT statement which fetches the values,
probably the other database adapters won't need these values.

The \cyclone\db\query\* classes all contain a method named prepare() which creates a prepared statement for the already built
SQL query. The prepare() implementations are now changed to pass $this to the newly created object representing the prepared
statement (see \cyclone\db\prepared\query\* classes). The prepared query (statemment) objects will all store the original
query in a private variable and their exec() method will pass this original query to the exec_* method of the
prepared executor implementation of the database adapter.